backend/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
numpy
faiss-cpu
sentence-transformers
pydantic-settings
